I've broken things badly in here, but MCK, SCK, and LRCK are right now.
Using the PCM5122 DAC, for 44100 Hz audio frequency out, MCK needs to be 128-256 times bigger.
So, setting the ACLK frequency to 256*44100 and ratio to 256 (which is forced at the moment) will leave the LRCK as 44100. Then, swidth can be set as desired (Using the NRF definitions now!) to allow SCK to be 2*LRCK*SWIDTH.

Going to clean this up later i think.